requirements.txt
paramiko>=3.3.1
requests>=2.31.0
ping3>=5.1.0
telethon>=1.36.0
python-telegram-bot>=20.7
pyvmomi>=8.0.3.0